refactor: move `instantiate_circuit` and `instantiate_empty_circuit` to utils
