+++ src/circuits/utils.rs
@@ -1,4 +1,6 @@
 use ark_std::{end_timer, start_timer};
+use crate::circuits::merkle_sum_tree::MerkleSumTreeCircuit;
+use crate::merkle_sum_tree::{big_int_to_fp, MerkleProof, MerkleSumTree};
 use halo2_proofs::{
     halo2curves::bn256::{Bn256, Fr as Fp, G1Affine},
     plonk::{create_proof, verify_proof, Circuit, ProvingKey, VerifyingKey},
@@ -15,6 +17,34 @@ use halo2_proofs::{
     },
 };
 use rand::rngs::OsRng;
+
+pub fn instantiate_circuit(assets_sum: Fp, path: &str) -> MerkleSumTreeCircuit {
+    let merkle_sum_tree = MerkleSumTree::new(path).unwrap();
+
+    let proof: MerkleProof = merkle_sum_tree.generate_proof(0).unwrap();
+
+    MerkleSumTreeCircuit {
+        leaf_hash: proof.entry.compute_leaf().hash,
+        leaf_balance: big_int_to_fp(proof.entry.balance()),
+        path_element_hashes: proof.sibling_hashes,
+        path_element_balances: proof.sibling_sums,
+        path_indices: proof.path_indices,
+        assets_sum,
+        root_hash: proof.root_hash,
+    }
+}
+
+pub fn instantiate_empty_circuit(levels: usize) -> MerkleSumTreeCircuit {
+    MerkleSumTreeCircuit {
+        leaf_hash: Fp::zero(),
+        leaf_balance: Fp::zero(),
+        path_element_hashes: vec![Fp::zero(); levels],
+        path_element_balances: vec![Fp::zero(); levels],
+        path_indices: vec![Fp::zero(); levels],
+        assets_sum: Fp::zero(),
+        root_hash: Fp::zero(),
+    }
+}
 
 pub fn full_prover<C: Circuit<Fp>>(
     params: &ParamsKZG<Bn256>,
